+++ modules/debugger/builtin/cpp/Gateway.cpp
@@ -1,0 +1,29 @@
+//=============================================================================
+// Copyright (c) 2016-present Allan CORNET (Nelson)
+//=============================================================================
+// This file is part of the Nelson.
+//=============================================================================
+// LICENCE_BLOCK_BEGIN
+// SPDX-License-Identifier: LGPL-3.0-or-later
+// LICENCE_BLOCK_END
+//=============================================================================
+#include "NelsonGateway.hpp"
+#include "debugger_Gateway.hpp"
+#include "dbstackBuiltin.hpp"
+//=============================================================================
+using namespace Nelson;
+//=============================================================================
+const std::wstring gatewayName = L"debugger";
+//=============================================================================
+static const nlsGateway gateway[] = {
+    { "dbstack", (ptrBuiltin)Nelson::DebuggerGateway::dbstackBuiltin, -1, -1,
+        CPP_BUILTIN_WITH_EVALUATOR },
+};
+//=============================================================================
+int
+DebuggerGateway(void* eval, const wchar_t* moduleFilename)
+{
+    return NelsonAddGatewayWithEvaluator(eval, moduleFilename, (void*)gateway,
+        sizeof(gateway) / sizeof(nlsGateway), gatewayName.c_str(), (void*)nullptr);
+}
+//=============================================================================