+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- debugger builtin added:
+  `dbstack`
+
 - integer builtin added:
   `int8`, `int16`
   `int32`, `int64`